Start splitting out the data model compare script.
This moves the main into a separate script, adds uv run support.
